move image upload button to a separated component
